Created Navigator page
Created navigator page for further listings
