+++ devvit.yaml
@@ -1,2 +1,2 @@
 name: gridlock-game2
-version: 0.0.2.331
+version: 0.0.2.338
